@@ -9,18 +9,21 @@ import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.inventory.Inventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
+import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.screen.slot.Slot;
 import net.minecraft.util.Identifier;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
-@Mixin(targets = {"net/minecraft/screen/PlayerScreenHandler$2"})
+import java.util.Optional;
+
+@Mixin(targets = {"net/minecraft/screen/PlayerScreenHandler$1"})
 public abstract class PlayerScreenHandlerOffHandSlotMixin extends Slot {
 
 	@Shadow
 	@Final
-	PlayerEntity field_42464;
+	PlayerEntity field_39410;
 
 	public PlayerScreenHandlerOffHandSlotMixin(Inventory inventory, int index, int x, int y) {
 		super(inventory, index, x, y);
@@ -29,16 +32,16 @@ public abstract class PlayerScreenHandlerOffHandSlotMixin extends Slot {
 	@Override
 	public boolean canInsert(ItemStack stack) {
 		boolean bl = true;
-		if (this.field_42464.getServer() != null) {
-			bl = this.field_42464.getServer().getGameRules().getBoolean(GameRulesRegistry.CAN_CHANGE_EQUIPMENT);
+		if (this.field_39410.getServer() != null) {
+			bl = this.field_39410.getServer().getGameRules().getBoolean(GameRulesRegistry.CAN_CHANGE_EQUIPMENT);
 		}
 
-		StatusEffect civilisation_status_effect = Registries.STATUS_EFFECT.get(Identifier.tryParse(RPGInventory.serverConfig.civilisation_status_effect_identifier));
-		boolean hasCivilisationEffect = civilisation_status_effect != null && this.field_42464.hasStatusEffect(civilisation_status_effect);
+		Optional<RegistryEntry.Reference<StatusEffect>> civilisation_status_effect = Registries.STATUS_EFFECT.getEntry(Identifier.tryParse(RPGInventory.serverConfig.civilisation_status_effect_identifier));
+		boolean hasCivilisationEffect = civilisation_status_effect.isPresent() && this.field_39410.hasStatusEffect(civilisation_status_effect.get());
 
-		StatusEffect wilderness_status_effect = Registries.STATUS_EFFECT.get(Identifier.tryParse(RPGInventory.serverConfig.wilderness_status_effect_identifier));
-		boolean hasWildernessEffect = wilderness_status_effect != null && this.field_42464.hasStatusEffect(wilderness_status_effect);
+		Optional<RegistryEntry.Reference<StatusEffect>> wilderness_status_effect = Registries.STATUS_EFFECT.getEntry(Identifier.tryParse(RPGInventory.serverConfig.wilderness_status_effect_identifier));
+		boolean hasWildernessEffect = wilderness_status_effect.isPresent() && this.field_39410.hasStatusEffect(wilderness_status_effect.get());
 
-		return stack.isIn(Tags.OFFHAND_ITEMS) && (hasCivilisationEffect || this.field_42464.isCreative() || (bl && !hasWildernessEffect)) && !((DuckPlayerEntityMixin) this.field_42464).rpginventory$isOffhandStackSheathed();
+		return stack.isIn(Tags.OFFHAND_ITEMS) && (hasCivilisationEffect || this.field_39410.isCreative() || (bl && !hasWildernessEffect)) && !((DuckPlayerEntityMixin) this.field_39410).rpginventory$isOffhandStackSheathed();
 	}
 }
