@@ -1,30 +1,23 @@
 package com.github.theredbrain.rpginventory.network.packet;
 
 import com.github.theredbrain.rpginventory.RPGInventory;
-import net.fabricmc.fabric.api.networking.v1.FabricPacket;
-import net.fabricmc.fabric.api.networking.v1.PacketType;
-import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.RegistryByteBuf;
+import net.minecraft.network.codec.PacketCodec;
+import net.minecraft.network.packet.CustomPayload;
 
-public class SheatheWeaponsPacket implements FabricPacket {
-	public static final PacketType<SheatheWeaponsPacket> TYPE = PacketType.create(
-			RPGInventory.identifier("sheathe_weapons"),
-			SheatheWeaponsPacket::new
-	);
+public record SheatheWeaponsPacket() implements CustomPayload {
+	public static final CustomPayload.Id<SheatheWeaponsPacket> PACKET_ID = new CustomPayload.Id<>(RPGInventory.identifier("sheathe_weapons"));
+	public static final PacketCodec<RegistryByteBuf, SheatheWeaponsPacket> PACKET_CODEC = PacketCodec.of(SheatheWeaponsPacket::write, SheatheWeaponsPacket::new);
 
-	public SheatheWeaponsPacket() {
-	}
-
-	public SheatheWeaponsPacket(PacketByteBuf buf) {
+	public SheatheWeaponsPacket(RegistryByteBuf registryByteBuf) {
 		this();
 	}
 
-	@Override
-	public PacketType<?> getType() {
-		return TYPE;
+	private void write(RegistryByteBuf registryByteBuf) {
 	}
 
 	@Override
-	public void write(PacketByteBuf buf) {
+	public CustomPayload.Id<? extends CustomPayload> getId() {
+		return PACKET_ID;
 	}
-
 }
