@@ -4,7 +4,6 @@ import com.github.theredbrain.rpginventory.RPGInventory;
 import com.github.theredbrain.rpginventory.entity.player.DuckPlayerEntityMixin;
 import com.github.theredbrain.rpginventory.entity.player.DuckPlayerInventoryMixin;
 import com.github.theredbrain.staminaattributes.entity.StaminaUsingEntity;
-import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -12,11 +11,12 @@ import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.text.Text;
 
-public class SwapHandItemsPacketReceiver implements ServerPlayNetworking.PlayPacketHandler<SwapHandItemsPacket> {
+public class SwapHandItemsPacketReceiver implements ServerPlayNetworking.PlayPayloadHandler<SwapHandItemsPacket> {
 	@Override
-	public void receive(SwapHandItemsPacket packet, ServerPlayerEntity player, PacketSender responseSender) {
+	public void receive(SwapHandItemsPacket payload, ServerPlayNetworking.Context context) {
 
-		boolean mainHand = packet.mainHand;
+		boolean mainHand = payload.mainHand();
+		ServerPlayerEntity player = context.player();
 
 		ItemStack itemStack;
 		ItemStack alternativeItemStack;
@@ -58,6 +58,6 @@ public class SwapHandItemsPacketReceiver implements ServerPlayNetworking.PlayPac
 		if (RPGInventory.isStaminaAttributesLoaded && !player.isCreative()) {
 			((StaminaUsingEntity) player).staminaattributes$addStamina(-RPGInventory.serverConfig.swapping_hand_items_stamina_cost);
 		}
-		player.getServerWorld().playSound(null, player.getBlockPos(), SoundEvents.ITEM_ARMOR_EQUIP_GENERIC, SoundCategory.PLAYERS, 1.0F, 1.0F);
+		player.getServerWorld().playSound(null, player.getBlockPos().getX(), player.getBlockPos().getY(), player.getBlockPos().getZ(), SoundEvents.ITEM_ARMOR_EQUIP_GENERIC, SoundCategory.PLAYERS, 1.0F, 1.0F);
 	}
 }
