@@ -5,7 +5,6 @@ import com.github.theredbrain.rpginventory.entity.player.DuckPlayerEntityMixin;
 import com.github.theredbrain.rpginventory.entity.player.DuckPlayerInventoryMixin;
 import com.github.theredbrain.rpginventory.registry.Tags;
 import com.github.theredbrain.staminaattributes.entity.StaminaUsingEntity;
-import net.fabricmc.fabric.api.networking.v1.PacketSender;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -13,9 +12,11 @@ import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvents;
 import net.minecraft.text.Text;
 
-public class ToggleTwoHandedStancePacketReceiver implements ServerPlayNetworking.PlayPacketHandler<ToggleTwoHandedStancePacket> {
+public class ToggleTwoHandedStancePacketReceiver implements ServerPlayNetworking.PlayPayloadHandler<ToggleTwoHandedStancePacket> {
 	@Override
-	public void receive(ToggleTwoHandedStancePacket packet, ServerPlayerEntity player, PacketSender responseSender) {
+	public void receive(ToggleTwoHandedStancePacket payload, ServerPlayNetworking.Context context) {
+
+		ServerPlayerEntity player = context.player();
 
 		ItemStack handItemStack = ((DuckPlayerInventoryMixin) player.getInventory()).rpginventory$getHand().copy();
 		if (handItemStack.isEmpty()) {
@@ -53,6 +54,6 @@ public class ToggleTwoHandedStancePacketReceiver implements ServerPlayNetworking
 		if (RPGInventory.isStaminaAttributesLoaded && !player.isCreative()) {
 			((StaminaUsingEntity) player).staminaattributes$addStamina(-RPGInventory.serverConfig.toggling_two_handed_stance_stamina_cost);
 		}
-		player.getServerWorld().playSound(null, player.getBlockPos(), SoundEvents.ITEM_ARMOR_EQUIP_GENERIC, SoundCategory.PLAYERS, 1.0F, 1.0F);
+		player.getServerWorld().playSound(null, player.getBlockPos().getX(), player.getBlockPos().getY(), player.getBlockPos().getZ(), SoundEvents.ITEM_ARMOR_EQUIP_GENERIC, SoundCategory.PLAYERS, 1.0F, 1.0F);
 	}
 }
