@@ -3,26 +3,27 @@ package com.github.theredbrain.rpginventory.registry;
 import com.github.theredbrain.rpginventory.RPGInventoryClient;
 import net.fabricmc.fabric.api.client.item.v1.ItemTooltipCallback;
 import net.minecraft.entity.EquipmentSlot;
-import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.item.Equipment;
 import net.minecraft.text.Text;
 
 public class ClientEventsRegistry {
 
 	public static void initializeClientEvents() {
-		ItemTooltipCallback.EVENT.register((stack, context, lines) -> {
+		ItemTooltipCallback.EVENT.register((stack, context, type, lines) -> {
 			if (stack.isIn(Tags.TWO_HANDED_ITEMS) && RPGInventoryClient.clientConfig.show_item_tooltip_two_handed_items) {
 				lines.add(Text.translatable("item.additional_tooltip.functionality.two_handed_item"));
 			}
 
 			// equipment slots
 			if (RPGInventoryClient.clientConfig.show_item_tooltip_equipment_slots) {
-				if (stack.isIn(Tags.HELMETS) || MobEntity.getPreferredEquipmentSlot(stack) == EquipmentSlot.HEAD) {
+				Equipment equipment = Equipment.fromStack(stack);
+				if (stack.isIn(Tags.HELMETS) || (equipment != null && equipment.getSlotType() == EquipmentSlot.HEAD)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.helmet"));
 				}
 				if (stack.isIn(Tags.NECKLACES)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.necklace"));
 				}
-				if (stack.isIn(Tags.CHEST_PLATES) || MobEntity.getPreferredEquipmentSlot(stack) == EquipmentSlot.CHEST) {
+				if (stack.isIn(Tags.CHEST_PLATES) || (equipment != null && equipment.getSlotType() == EquipmentSlot.CHEST)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.chest_plate"));
 				}
 				if (stack.isIn(Tags.SHOULDERS)) {
@@ -51,10 +52,10 @@ public class ClientEventsRegistry {
 				if (stack.isIn(Tags.SPELLS)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.spell"));
 				}
-				if (stack.isIn(Tags.LEGGINGS) || MobEntity.getPreferredEquipmentSlot(stack) == EquipmentSlot.LEGS) {
+				if (stack.isIn(Tags.LEGGINGS) || (equipment != null && equipment.getSlotType() == EquipmentSlot.LEGS)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.leggings"));
 				}
-				if (stack.isIn(Tags.BOOTS) || MobEntity.getPreferredEquipmentSlot(stack) == EquipmentSlot.FEET) {
+				if (stack.isIn(Tags.BOOTS) || (equipment != null && equipment.getSlotType() == EquipmentSlot.FEET)) {
 					lines.add(Text.translatable("item.additional_tooltip.equipment_slot.boots"));
 				}
 			}
